Refactor: 코드 리뷰 반영 #71
