+++ src/components/MyPage/Reservations/Details/CheckStatus.module.css
@@ -1,0 +1,25 @@
+.status {
+  display: flex;
+  justify-self: flex-start;
+  align-items: center;
+  font-size: 1.8rem;
+  font-weight: 700;
+  line-height: 2.6rem;
+  gap: 0.5rem;
+}
+
+.active {
+  color: var(--blue2E);
+}
+
+.inactive {
+  color: var(--grayA1);
+}
+
+.canceled {
+  color: var(--gray79);
+}
+
+.declined {
+  color: var(--redFF4);
+}
