+++ src/components/MyPage/Reservations/Details/CheckStatus.tsx
@@ -1,6 +1,6 @@
 import ArrowRightIcon from '#/icons/icon-arrowRight.svg';
 import { RESERVATION_STATUS, ReservationStatus } from '@/constants/reservation';
-import styles from '@/pages/mypage/reservations/[id]/ReservationId.module.css';
+import styles from './CheckStatus.module.css';
 
 interface CheckStatusProps {
   status: keyof typeof ReservationStatus;
@@ -10,7 +10,7 @@ const CheckStatus = ({ status }: CheckStatusProps) => {
   return (
     <div className={styles.status}>
       {status === RESERVATION_STATUS.CANCELED || status === RESERVATION_STATUS.DECLINED ? (
-        <div className={styles[status]}>{ReservationStatus[status as keyof typeof ReservationStatus]}</div>
+        <div className={styles[status]}>{ReservationStatus[status]}</div>
       ) : (
         <>
           <div className={status === RESERVATION_STATUS.PENDING ? styles.active : styles.inactive}>
